[IMP] project(_todo): generic improvement for to-do app
In this commit, did the below changes:
- Remove the possibility of adding/editing/removing assignees from the
  kanban view.
- Make the input for the title bigger in the to-do form view.
- When we add a new line, it will focus on the title field.
- In the to-dos list view make the assignee's field required.
- Reduce the size of the 'convert to task' wizard.
- The 'convert to task' wizard removed the placeholder of the project field.
- When a to-do is marked as done, the button marks it as done then the
  tooltip is done.

task-3931142
