+++ addons/project/static/src/components/subtask_one2many_field/subtask_list_renderer.js
@@ -1,29 +1,10 @@
 /** @odoo-module **/
 
 import { _t } from "@web/core/l10n/translation";
-import { useService } from "@web/core/utils/hooks";
 import { ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
-import { ListRenderer } from '@web/views/list/list_renderer';
+import { TaskListRenderer } from "../task_list_renderer";
 
-import { useEffect } from "@odoo/owl";
-
-export class SubtaskListRenderer extends ListRenderer {
-    setup() {
-        super.setup();
-        this.dialog = useService("dialog");
-        useEffect(
-            () => this.focusName(this.props.list.editedRecord),
-            () => [this.props.list.editedRecord]
-        );
-    }
-
-    focusName(editedRecord) {
-        if (editedRecord?.isNew && !editedRecord.dirty) {
-            const col = this.columns.find((c) => c.name === "name");
-            this.focusCell(col);
-        }
-    }
-
+export class SubtaskListRenderer extends TaskListRenderer {
     async onDeleteRecord(record) {
         this.dialog.add(ConfirmationDialog, {
             body: _t("Are you sure you want to delete this record?"),
