+++ addons/project/static/src/components/task_list_renderer.js
@@ -1,0 +1,22 @@
+import { useService } from "@web/core/utils/hooks";
+import { ListRenderer } from "@web/views/list/list_renderer";
+
+import { useEffect } from "@odoo/owl";
+
+export class TaskListRenderer extends ListRenderer {
+    setup() {
+        super.setup();
+        this.dialog = useService("dialog");
+        useEffect(
+            () => this.focusName(this.props.list.editedRecord),
+            () => [this.props.list.editedRecord]
+        );
+    }
+
+    focusName(editedRecord) {
+        if (editedRecord?.isNew && !editedRecord.dirty) {
+            const col = this.columns.find((c) => c.name === "name");
+            this.focusCell(col);
+        }
+    }
+}
