+++ addons/project_todo/static/src/components/todo_done_checkmark/todo_done_checkmark.xml
@@ -3,14 +3,14 @@
 
     <t t-name="project_todo.TodoDoneCheckmark">
         <t t-if="!env.isSmall">
-            <a title="Mark as done"
+            <a t-att-title="stateDone.isDone ? 'Mark as to-do' : 'Mark as done'"
                t-on-click.stop="onDoneToggled"
                t-on-mouseleave="actualizeDoneState"
                t-on-mouseover="freezeDoneState"
                t-attf-class="o_todo_done_button fa fa-lg fa-check-circle{{!stateDone.isDone ? '-o' : ' done_button_enabled'}}"/>
         </t>
         <t t-else="">
-            <a title="Mark as done"
+            <a t-att-title="stateDone.isDone ? 'Mark as to-do' : 'Mark as done'"
                t-on-click.stop="onDoneToggled"
                t-attf-class="o_todo_done_button_mobile fa fa-lg fa-check-circle{{!stateDone.isDone ? '-o' : ' done_button_enabled'}}"/>
         </t>
