+++ addons/project_todo/static/src/views/todo_list/todo_list_view.js
@@ -3,10 +3,12 @@
 import { registry } from "@web/core/registry";
 import { listView } from "@web/views/list/list_view";
 import { TodoListController } from "./todo_list_controller";
+import { TaskListRenderer } from "@project/components/task_list_renderer";
 
 export const todoListView = {
     ...listView,
     Controller: TodoListController,
+    Renderer: TaskListRenderer,
 };
 
 registry.category("views").add("todo_list", todoListView);
